# openapi.yml
openapi: 3.0.3
info:
  title: Property Billing API
  version: 0.1.0
  description: API for apartment and housing monthly fee tracking.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server.
...
paths:
...
  /properties:
    get:
      tags:
        - Properties
      summary: List properties.
      description: Returns properties filtered by optional search text and status.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Status"
      responses:
        "200":
          description: Properties were successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyIndexResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - Properties
      summary: Create a property.
      description: Creates a property that can contain multiple units.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyCreateRequest"
      responses:
        "201":
          description: Property was created.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /properties/{property_id}:
    delete:
      tags:
        - Properties
      summary: Deactivate a property.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags:
        - Properties
      summary: Get property detail.
      description: Returns one property by ID.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      responses:
        "200":
          description: Property detail was successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyShowResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags:
        - Properties
      summary: Update a property.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdateRequest"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 100
    PropertyId:
      name: property_id
      in: path
      required: true
      description: UUID of the property.
      schema:
        type: string
        format: uuid
...
    Search:
      name: search
      in: query
      required: false
      description: Search keyword.
      schema:
        type: string
        maxLength: 100
    Status:
      name: status
      in: query
      required: false
      description: Property status filter.
      schema:
        $ref: "#/components/schemas/PropertyStatus"
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip before returning results.
      schema:
        type: integer
        minimum: 0
        default: 0
...
  responses:
    BadRequest:
      description: The request is invalid.
...
    NoContent:
      description: The request was successful and no content was returned.
    NotFound:
      description: The requested resource was not found.
    Unauthorized:
      description: Authentication is required and has failed.
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PropertyCreateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
          pattern: '.*\S.*'
          example: Green Residence
        address:
          type: string
          nullable: true
          maxLength: 500
          example: Bekasi
      required:
        - name
...
    PropertyIndexResponse:
      type: object
      properties:
        count:
          type: integer
          minimum: 0
          description: Number of items returned in `properties`.
          example: 1
        properties:
          type: array
          items:
            $ref: "#/components/schemas/PropertyIndexElement"
      required:
        - count
        - properties
    PropertyIndexElement:
      allOf:
        - $ref: "#/components/schemas/PropertyShowResponse"
    PropertyShowResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Green Residence
        address:
          type: string
          nullable: true
          example: Bekasi
        active:
          type: boolean
          example: true
      required:
        - id
        - name
        - active
    PropertyStatus:
      type: string
      enum:
        - active
        - inactive
      example: active
...
    PropertyUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/PropertyCreateRequest"
...